refactor: Handler는 단순 웹소켓 생명주기 관리로 축소, 메시지 처리는 Dispatcher로 위임
